backend/pyproject.toml
[tool.poetry]
name = "edge-backend"
version = "0.1.0"
description = "Edge backend scaffold"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
PyJWT = "^2.8.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
httpx = "^0.24.0"
ruff = "^0.12.0"
black = "^24.0"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"

[tool.ruff]
line-length = 88
